requirements.txt
numpy
torch
transformers
sentence-transformers
faiss-cpu
tqdm
scikit-learn
python-dotenv
